+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog for *TopoBank*
 
-## 1.54.0 (2025-02-04)
+## 1.54.0 (2025-02-07)
 
 - MAINT: Changed URL for getting names of properties in tags
 - BUG: Replace `login_required` by appropriate DRF authentication decorator
 - BUG: Return 404 when tag does not exist
+- BUG: Use `PermissionDenied` rather than `PermissionError` so that permission
+  errors are reported to the user
 
 ## 1.53.3 (2025-02-02)